requirements.txt
streamlit
Pillow
torch
transformers
datasets
tqdm